+++ src/app/(domains)/(home)/_sections/Footer.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ROUTES } from "@/constants/Routes";
+import { ROUTES } from "@/constants/ROUTES";
 import { useI18n } from "@/hooks/useI18n"
 
 type FooterProps = {
