+++ src/constants/FIXED_CURRENCY.ts
@@ -1,0 +1,1 @@
+export const FIXED_CURRENCY = 'usd';
