+++ src/constants/HAS_FREE_TRIAL.ts
