+++ src/constants/PASSWORD_STRENGTH_OPTIONS.ts
